pyproject.toml
[project]
name = "miku"
version = "0.1.0"
description = "a quirky discord bot"
authors = [
    {name = "Max!!",email = "dizziedbliss@gmail.com"}
]
license = {text = "no one cares"}
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "hikari (>=2.5.0,<3.0.0)",
    "hikari-lightbulb (>=3.2.3,<4.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "orjson (>=3.11.6,<4.0.0)",
    "yarl (>=1.22.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
